Add BigInt

+++ src/commonMain/kotlin/presentation/adapters/ScoreTextAdapter.kt
@@ -1,8 +1,10 @@
 package presentation.adapters
 
+import com.soywiz.kbignum.*
+
 object ScoreTextAdapter {
 
-    fun getTextByScore(score: Int): String {
+    fun getTextByScore(score: BigInt): String {
         return BlockTextAdapter.generateTextByStringValue(score.toString())
     }
 }
